restructuring parts of the code
