+++ src/main/java/edu/ie3/netpad/tool/grid/LineLengthResolutionMode.java
@@ -3,7 +3,7 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
-package edu.ie3.netpad.tool;
+package edu.ie3.netpad.tool.grid;
 
 public enum LineLengthResolutionMode {
   GEOGRAPHICAL,
